Right Controller Logic Separated
Separated Ray, Direct, and Transport Interactors on the right controller and used a system to flip between necessary logic when the user presses the appropriate inputs.

Controller grips and initial ray casting will only be used to interact with non transport interactables. When the player presses the right joy stick forward (north), transport logic is activated.